set

set

set

Author

Joe DeRose and Rick Mortensen

Rick Mortensen

Methods

Public methods

Junction$new()

Junction$removeDups()

Junction$gw()

Junction$subset()

Junction$set()

Junction$print()

Junction$clone()

Method new()

Builds junction class, grl must be GRangesList with each GRanges of length 2 JUNCTION Empty junctions are removed If grl is empty GRangesList, Junctions class is empty

Usage

Junction$new(grl = NULL, ...)

Arguments

grl

GRangesList of signed breakpoint pairs

Method removeDups()

Removes duplicate junctions in this Junction Object

this Junction Object has no duplicate junctions

Usage

Junction$removeDups()

Method gw()

Returns a gWalk representing the left -> right traversal of this junction +- padding

Usage

Junction$gw(pad = 0)

Returns

GRangesList of the junctions in this Junction Object

Method subset()

Allows subseting of the Junction object using bracket notation

Usage

Junction$subset(i)

Arguments

i

integer or self$length logical vector specifying subset

Method set()

set metadata for junction object

Usage

Junction$set(...)

Details

sets metadata to either a scalar or vector where the vector is the same length as the junction object ``` jj$set(cn = 7) jj$set(cn = c(1,2,3)) jj$set(col = c("red", "blue", "green")) ``` (after setting, metadata is accessible through $dt accessor)

Method print()

Prints out the Junction Object. Prints the length and the GRangesList of the junctions.

Usage

Junction$print()

Method clone()

The objects of this class are cloneable with this method.

Usage

Junction$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.