work horse graph function

work horse graph function

work horse graph function

Methods

Public methods

PGVdb$new()

PGVdb$addgraphs()

PGVdb$push()

PGVdb$refreshPGV()

PGVdb$clone()

Method new()

Usage

PGVdb$new(datafiles.json = NA, datafolder = NA, PGV_public_dir = NA)

Method addgraphs()

Usage

PGVdb$addgraphs(self, PGVInput, cores = 10)

Arguments

PGVInput

PGVInput class

cores

number of cores to use push object

Details

accepts PGVInput class and pushes it

Method push()

Usage

PGVdb$push(self, backup = TRUE)

Details

pushes object

Refresh object

Method refreshPGV()

Usage

PGVdb$refreshPGV(self)

Details

refreshes object from source given self. Used for any changes since the object was deserialized

Method clone()

The objects of this class are cloneable with this method.

Usage

PGVdb$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.