Amplifications

Classifies high-level amplifications in copy number annotated gGraph. Default parameters should be used.

amp(
  gg,
  jcn.thresh = 8,
  cn.thresh = 2,
  fbi.cn.thresh = 0.5,
  n.jun.high.bfb.thresh = 26,
  n.jun.high.dm.thresh = 31,
  width.thresh = 1e+05,
  fbi.width.thresh = 1e+05,
  mc.cores = 1,
  mark = TRUE,
  mark.col = "purple",
  mark.nos = FALSE,
  min.nodes = 3,
  min.jun = 2
)

Arguments

gg

gGraph

jcn.thresh

minimal ALT edge junction threshold to classify a high copy cluster. Default: 9

cn.thresh

minimal node copy number in ploidy units to classify a high copy cluster. Default: 2

fbi.cn.thresh

fraction of total CN in cluster that is contributed to by fold back inversions, if higher than this will call a BFB. Default: 0.5

n.jun.high.bfb.thresh

max number of high copy junctions in a fbi.cn high cluster, if fbi.cn is high and high copy junctions exceed this, then will call a tyfonas. Default: 26

n.jun.high.dm.thresh

double minute threshold set for junctions. If fbi.cn is low, and high copy junctions exceed this thresh, it will call it cpxdm, else dm. Default: 31

width.thresh

minimum width to consider for an amplification event. Default: 1e5

mark.nos

(logical) Default: FALSE

min.nodes

(numeric) minimum number of nodes for a cluster to be designated amp-NOS. Default: 3

min.jun

(numeric) minimum number of aberrant junctions for a cluster to be designated amp-NOS. Default: 2

Value

gg of amplification events found.

Details

Amplification events are defined into 3 groups, bfb, tyfonas, and dm events. More details can be found in the tutorial: Complex Amplicons