Rcplex2

Modification of Rcplex which takes in mipcontrol parameters.

Call simple events in gGraph

Rcplex2(
  cvec,
  Amat,
  bvec,
  Qmat = NULL,
  lb = 0,
  ub = Inf,
  control = list(),
  objsense = c("min", "max"),
  sense = "L",
  vtype = NULL,
  n = 1,
  tuning = FALSE
)

simple(
  gg,
  reciprocal.thresh = 10000,
  tra.pad = 1e+06,
  mark = TRUE,
  mark.col = "purple"
)

Arguments

gg

gGraph, must have 'cn' node and edge annotation

mark

logical flag specifying whether to mark fold backs (TRUE)

mark.col

character specifying colors to mark graph with ('purple')

min.foldbacks

minimal number of foldbacks that makes a bfb

amp.thresh

amplification threshold for a BFB

cn.bfb

minimal number of copies that the highest copy foldback in the bfb needs to have

Value

gGraph object containing labeling the putative event