merge junctions by overlaps with padding Merges a set of junctions and keeps "seen.by" metadata of junction origin using the argument names to this function If cartesian = TRUE, can only merge a pair of junction objects but then the output contains the overlapping junctions and metadata annotated with a $query.id (index into first argument) and $subject.id (index to into second argument) with deduped outputs,

merge junctions by overlaps with padding

Merges a set of junctions and keeps "seen.by" metadata of junction origin using the argument names to this function

If cartesian = TRUE, can only merge a pair of junction objects but then the output contains the overlapping junctions and metadata annotated with a $query.id (index into first argument) and $subject.id (index to into second argument) with deduped outputs,

# S3 method for Junction
merge(
  ...,
  pad = 0,
  ind = FALSE,
  cartesian = FALSE,
  all = FALSE,
  all.x = all,
  all.y = all
)

Arguments

...

GRangesList representing rearrangements to be merged

pad

non-negative integer specifying padding

ind

logical flag (default FALSE) specifying whether the "seen.by" fields should contain indices of inputs (rather than logical flags) and NA if the given junction is missing

cartesian

whether to do a pairwise merge of all junction pairs in two junction objects x and y, which can potentially result in more rows than the number of inputs, Note: only works when there are exactly two inputs x and y

all

only applicable if cartesian = TRUE, logical flag specifying whether to keep the junctions and metadata for non-overlapping junction pairs from both x and y inputs, aka "outer join" + "inner join"

all.x

only applicable if cartesian = TRUE, logical flag specifying whether to keep the junctions and metadata for non-overlapping junction pairs from j1 in the output aka "left join" + "inner join""

all.y

only applicable if cartesian = TRUE, logical flag specifying whether to keep the junctions and metadata for non-overlapping junction pairs from j2 in the output aka "right join" + "inner join"

Examples

## wil output a Junction object with metadata seen.by.svaba etc.
## will pad with 500 bases prior to merging

svaba = jJ(system.file('extdata', "HCC1143.svaba.somatic.sv.vcf", package = "gGnome"))
delly = jJ(system.file('extdata', "delly.final.vcf.gz", package = "gGnome"))
novobreak = jJ(system.file('extdata', "novoBreak.pass.flt.vcf", package = "gGnome"))
#> Warning: Some breakpoint width==0.
#> Warning: Warning: Check or fix seqlengths, some are equal 0 or NA, may lead to negative widths
#> Warning: GRanges object contains 2 out-of-bound ranges located on sequences 16
#>   and Y. Note that ranges located on a sequence whose length is unknown
#>   (NA) or on a circular sequence are not considered out-of-bound (use
#>   seqlengths() and isCircular() to get the lengths and circularity flags
#>   of the underlying sequences). You can use trim() to trim these ranges.
#>   See ?`trim,GenomicRanges-method` for more information.
#> Warning: GRanges object contains 2 out-of-bound ranges located on sequences 16
#>   and Y. Note that ranges located on a sequence whose length is unknown
#>   (NA) or on a circular sequence are not considered out-of-bound (use
#>   seqlengths() and isCircular() to get the lengths and circularity flags
#>   of the underlying sequences). You can use trim() to trim these ranges.
#>   See ?`trim,GenomicRanges-method` for more information.
#> Warning: GRanges object contains 2 out-of-bound ranges located on sequences 16
#>   and Y. Note that ranges located on a sequence whose length is unknown
#>   (NA) or on a circular sequence are not considered out-of-bound (use
#>   seqlengths() and isCircular() to get the lengths and circularity flags
#>   of the underlying sequences). You can use trim() to trim these ranges.
#>   See ?`trim,GenomicRanges-method` for more information.
#> Warning: GRanges object contains 2 out-of-bound ranges located on sequences 16
#>   and Y. Note that ranges located on a sequence whose length is unknown
#>   (NA) or on a circular sequence are not considered out-of-bound (use
#>   seqlengths() and isCircular() to get the lengths and circularity flags
#>   of the underlying sequences). You can use trim() to trim these ranges.
#>   See ?`trim,GenomicRanges-method` for more information.
#> Warning: GRanges object contains 2 out-of-bound ranges located on sequences 16
#>   and Y. Note that ranges located on a sequence whose length is unknown
#>   (NA) or on a circular sequence are not considered out-of-bound (use
#>   seqlengths() and isCircular() to get the lengths and circularity flags
#>   of the underlying sequences). You can use trim() to trim these ranges.
#>   See ?`trim,GenomicRanges-method` for more information.
#> Warning: GRanges object contains 2 out-of-bound ranges located on sequences 16
#>   and Y. Note that ranges located on a sequence whose length is unknown
#>   (NA) or on a circular sequence are not considered out-of-bound (use
#>   seqlengths() and isCircular() to get the lengths and circularity flags
#>   of the underlying sequences). You can use trim() to trim these ranges.
#>   See ?`trim,GenomicRanges-method` for more information.
#> Warning: GRanges object contains 2 out-of-bound ranges located on sequences 16
#>   and Y. Note that ranges located on a sequence whose length is unknown
#>   (NA) or on a circular sequence are not considered out-of-bound (use
#>   seqlengths() and isCircular() to get the lengths and circularity flags
#>   of the underlying sequences). You can use trim() to trim these ranges.
#>   See ?`trim,GenomicRanges-method` for more information.
#> Warning: GRanges object contains 2 out-of-bound ranges located on sequences 16
#>   and Y. Note that ranges located on a sequence whose length is unknown
#>   (NA) or on a circular sequence are not considered out-of-bound (use
#>   seqlengths() and isCircular() to get the lengths and circularity flags
#>   of the underlying sequences). You can use trim() to trim these ranges.
#>   See ?`trim,GenomicRanges-method` for more information.
#> Warning: GRanges object contains 2 out-of-bound ranges located on sequences 16
#>   and Y. Note that ranges located on a sequence whose length is unknown
#>   (NA) or on a circular sequence are not considered out-of-bound (use
#>   seqlengths() and isCircular() to get the lengths and circularity flags
#>   of the underlying sequences). You can use trim() to trim these ranges.
#>   See ?`trim,GenomicRanges-method` for more information.
#> Warning: GRanges object contains 2 out-of-bound ranges located on sequences 16
#>   and Y. Note that ranges located on a sequence whose length is unknown
#>   (NA) or on a circular sequence are not considered out-of-bound (use
#>   seqlengths() and isCircular() to get the lengths and circularity flags
#>   of the underlying sequences). You can use trim() to trim these ranges.
#>   See ?`trim,GenomicRanges-method` for more information.

## merge(svaba = svaba, delly = delly, caller3 = novobreak, pad = 500)