phased.binstats

Junction-balanced version of jabba.alleles

Given a balanced unphased genome graph with field $cn populated with integer values Constructs a "melted" haplotype graph with fields $cn and $allele

Given GRanges containing major/minor allele counts and a balanced but unphased gGraph, prepares phased gGraph input to balance.

phased.binstats.legacy(
  gg,
  bins = NULL,
  purity = NULL,
  ploidy = NULL,
  count.field = "count",
  allele.field = "allele",
  phase.blocks = NULL,
  edge.phase.dt = NULL,
  vbase.count.thres = 5,
  vbase.prop.thres = 0.9,
  min.bins = 3,
  min.var = 0.001,
  verbose = TRUE,
  mc.cores = 8
)

Arguments

gg

gGraph

bins

GRanges with:

purity

(numeric)

ploidy

(numeric)

count.field

(str) field containing allele read counts (default count)

allele.field

(str) field for containing allele label for read counts (default allele)

phase.blocks

(GRanges) GRanges containing phase blocks (e.g. from HAPCUT). default NULL.

edge.phase.dt

(data.table) with columns n1.major, n2.major, n1.minor, n2.minor and edge.id providing major/minor allele counts

vbase.count.thres

(int) number of variant base counts required to phase edges (default 5)

vbase.prop.thres

(float) proportion of allele excess required to phase edges (default 0.9)

min.bins

(numeric) minimum number of bins for intra segment variance (default 3)

min.var

(numeric) min allowable variance (default 0.1)

verbose

(bool) default TRUE for debugging

mc.cores

(int) number of cores

Value

gGraph whose nodes are annotated with $cn.major, $cn.minor, $haplotype, and $weight fields