read.juncs

load SV data as GRangesList

read.juncs(
  rafile,
  keep.features = TRUE,
  seqlengths = NULL,
  chr.convert = FALSE,
  get.loose = FALSE,
  standard.only = FALSE,
  flipstrand = FALSE,
  verbose = FALSE,
  ...
)

Arguments

rafile

(character) path to junctions file

keep.features

(logical) keep metadata? default TRUE

seqlengths

(numeric) a named numeric vector of contig lengths. default NULL

chr.convert

(logical) strip chr prefix on contig names? default FALSE

get.loose

(logical) get loose ends. warning: not implemented yet!

standard.only

(logical) retain only junctions between standard assembled chromosomes. default FALSE

flipstrand

(logical) flip junction strands? default FALSE

verbose

(logical) default FALSE

Value

GRangesList if get.loose = FALSE list if get.loose = TRUE (with slots $junctions and $loose) where $junctions is GRangesList and $loose is a signed GRanges

Details

Reads a file containing SVs and returns a GRangesList. Supported file types are: - .rds (containing GRangesList or Junction object) - .bedpe/.bedpe.gz - .vcf/.vcf.gz

VCF files produced from the following WGS junction callers are supported: - SvABA (https://github.com/walaj/svaba) - GRIDSS (https://github.com/PapenfussLab/gridss) - delly (https://github.com/dellytools/delly) - novoBreak (https://github.com/czc/nb_distribution)

In addition, the following long read callers are supported: - SVIM (https://github.com/eldariont/svim/wiki) - PBSV (https://github.com/PacificBiosciences/pbsv) - Sniffles2 (https://github.com/fritzsedlazeck/Sniffles) - cuteSV (https://github.com/tjiangHIT/cuteSV)

This function will load rearangements with the following SVTYPE values: - DEL (requires END field) - DUP (requires END field) - INV (requires END field) - TRA (for this SVTYPE, the INFO fields CHR2 and CT are also required) - BND (requires ALT field with proper formatting)

Note that we currently ignore SVTYPE INS rearrangements. In addition, SVTYPE BND rearrangements must have an ALT field corresponding to a single distal site. At this time, we will not load single breakends without a distal site, or breakends with multiple distal sites.