rel2abs

rescales CN values from relative to "absolute" (i.e. per cancer cell copy) scale given purity and ploidy

rel2abs(
  gr,
  purity = NA,
  ploidy = NA,
  gamma = NA,
  beta = NA,
  field = "ratio",
  field.ncn = "ncn"
)

Arguments

gr

GRanges input with meta data field corresponding to mean relative copy "mean" in that interval

purity

purity of sample

ploidy

ploidy of sample

gamma

gamma fit of solution (over-rides purity and ploidy)

beta

beta fit of solution (over-rides purity and ploidy)

field

meta data field in "gr" variable from which to extract signal, default "mean"

field.ncn

meta data field in "gr" variable from which to extract germline integer copy number, default "ncn", if doesn't exist, germline copy number is assumed to be zero

Value

numeric vector of integer copy numbers

Details

takes in gr with signal field "field"