Find Chromoplexy chains

Finds chromoplexy chains as clusters of "long distance" junctions that each span at least min.span (i.e. distant regions on the reference) have junctions nearby ie within max.dist. We filter to chains that have at least min.num footprints on the genome and involve at least min.num long distance junctions. We keep track of how many "other" (non small dup and non small del) junctions there are in the vicinity for downstream filtering.

chromoplexy(
  gg,
  min.span = 1e+07,
  max.dist = 10000,
  min.num = 3,
  max.cn = 3,
  footprint.width = 1e+06,
  ignore.small.dups = TRUE,
  ignore.small.dels = TRUE,
  max.small = 50000,
  mark = FALSE,
  mark.col = "purple"
)

Arguments

gg

gGraph

min.span

minimimum span to define a "long distance" junction and also the span by which major footprints of the event must be separated Default: 1e7

max.dist

maximum distance allowed in edge clusters Default: 1e4

min.num

minimum number of junctions and major footprints that define a chromoplexy Default: 3

max.cn

max copy number before filter Default: 3

footprint.width

padding around which to define the footprint of an event, note that the outputted footprint only includes the chromoplexy junction breakpoints. Default: 1e6

ignore.small.dups

logical flag determining whether we ignore small dups when filtering on min.cushion. Default: True

ignore.small.dels

logical flag determining whether we ignore small dels when filtering on min.cushion. Default: True

max.small

threshold for calling a local dup or del "small". Default: 5e4

mark

mark chromoplexies. Default:FALSE

mark.col

color to mark chromoplexies. Default: purple

Value

gGraph with $meta$chromoplexy annotated with chromoplexy event metadata and edges labeled with $chromoplexy id or NA if the edge does not belong to a chromoplexy

Details

For more details on what chromoplexies are and visualized: Chromoplexy and TICs